+++ frontend/src/components/Metrics/ConfigStep/DateRangePicker/index.tsx
@@ -4,7 +4,7 @@ import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider'
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs'
 import { DatePicker } from '@mui/x-date-pickers'
 import { TextField } from '@mui/material'
-import { datePickerPropsStyles, DateRangeBox } from '@src/components/metrics/ConfigStep/DateRangePicker/style'
+import { datePickerPropsStyles, DateRangeBox } from '@src/components/Metrics/ConfigStep/DateRangePicker/style'
 import { DATE_RANGE, SELECT_OR_WRITE_DATE } from '@src/constants'
 import { useAppDispatch, useAppSelector } from '@src/hooks'
 import { changeBoardVerifyState } from '@src/features/board/boardSlice'
