+++ frontend/src/components/Metrics/ConfigStep/MetricsTypeCheckbox/index.tsx
@@ -1,8 +1,8 @@
 import { Checkbox, FormHelperText, InputLabel, ListItemText, MenuItem, Select, SelectChangeEvent } from '@mui/material'
 import { REQUIRED_DATAS } from '@src/constants'
 import React, { useState } from 'react'
-import { RequireDataSelections } from '@src/components/metrics/ConfigStep/MetricsTypeCheckbox/style'
-import { Board } from '@src/components/metrics/ConfigStep/Board'
+import { RequireDataSelections } from '@src/components/Metrics/ConfigStep/MetricsTypeCheckbox/style'
+import { Board } from '@src/components/Metrics/ConfigStep/Board'
 import { useAppDispatch, useAppSelector } from '@src/hooks'
 import { selectRequiredData, updateRequiredData } from '@src/features/config/configSlice'
